Move everything into form
