+++ packages/desktop/app/components/index.js
@@ -14,6 +14,7 @@ export * from './ModalActionRow';
 export * from './MenuButton';
 export * from './Notification';
 export * from './PageContainer';
+export * from './PatientDetailsCard';
 export * from './Paper';
 export * from './Preloader';
 export * from './LocationSearchBar';
